requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.109.2
googlemaps==4.10.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
joblib==1.3.2
Mako==1.3.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==2.6.1
pydantic_core==2.16.2
python-jose==3.3.0
python-multipart==0.0.9
regex==2023.12.25
requests==2.31.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
threadpoolctl==3.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
